Move operation rate-limit bucket cleaner next to the other clean-up jobs
Relocate OperationRateLimitBucketCleaner into the process module's jobs
package, alongside the existing clean-up jobs, and move its BucketStore /
PostgresBucketStore collaborators into the process util package. The
bucket4j dependency, the Postgres integration test, and the failsafe
plugin move to the process module accordingly; the web limiter now uses
the bucket store transitively. The three cleaner log messages move to the
process Messages class. No behavior change.
