+++ multiapps-controller-process/src/main/java/module-info.java
@@ -67,5 +67,7 @@ open module org.cloudfoundry.multiapps.controller.process {
     requires static java.compiler;
     requires static org.immutables.value;
     requires org.cloudfoundry.multiapps.controller.shutdown.client;
+    requires io.github.bucket4j.core;
+    requires io.github.bucket4j.postgresql;
 
 }
