+++ multiapps-controller-web/src/main/java/org/cloudfoundry/multiapps/controller/web/util/OperationRateLimiter.java
@@ -7,6 +7,7 @@ import jakarta.inject.Named;
 
 import org.cloudfoundry.multiapps.controller.core.util.ApplicationConfiguration;
 import org.cloudfoundry.multiapps.controller.persistence.services.OperationService;
+import org.cloudfoundry.multiapps.controller.process.util.BucketStore;
 import org.cloudfoundry.multiapps.controller.web.Messages;
 
 import io.github.bucket4j.Bandwidth;
